+++ app/controllers/concerns/dradis/plugins/persistent_permissions.rb
@@ -3,13 +3,15 @@ module Dradis
     module PersistentPermissions
       extend ActiveSupport::Concern
 
+      include UsageTracking
+
       def update
         @user = User.authors.find(params[:id])
 
         Permission.transaction do
           Permission.where(component: self.class.component_name, user_id: params[:id]).destroy_all
 
-          permissions_params[:permissions]&.each do |permission|
+          permission_params[:permissions]&.each do |permission|
             # Validate the permission being created is a valid value
             next unless self.class.permissions_validation.call(permission) if self.class.permissions_validation
 
@@ -21,12 +23,18 @@ module Dradis
           end
         end
 
+        track_usage(event_name, { id: params[:id], params: permission_params })
+
         redirect_to main_app.edit_admin_user_permissions_path(params[:id]), notice: "#{@user.name}'s permissions have been updated."
       end
 
       private
 
-      def permissions_params
+      def event_name
+        "#{self.class.component_name}_permissions.updated"
+      end
+
+      def permission_params
         params.require(self.class.component_name).permit(permissions: [])
       end