+++ lib/dradis/plugins/gem_version.rb
@@ -7,8 +7,8 @@ module Dradis
 
     module VERSION
       MAJOR = 4
-      MINOR = 11
-      TINY  = 0
+      MINOR = 15
+      TINY = 0
       PRE   = nil
 
       STRING = [MAJOR, MINOR, TINY, PRE].compact.join('.')