+++ CHANGELOG.md
@@ -1,6 +1,20 @@
-v4.12.0 (XXXX 2024)
+v4.16.0 (Month 2025)
+  - Enable audit tracking for persistent permissions changes
+
+v4.15.0 (December 2024)
+  - No changes
+
+v4.14.0 (October 2024)
+  - No changes
+
+v4.13.0 (July 2024)
+  - No changes
+
+v4.12.0 (May 2024)
   - Update Dradis links in README
   - Fix the TypeError around the plugins template caching
+  - Remove template_service and add mapping_service to apply mappings on tool upload
+  - Include mapping module when integrations provide 'upload'
 
 v4.11.0 (January 2024)
   - No changes